cmdb/requirements.txt
amqp==2.0.3
anyjson==0.3.3
billiard==3.5.0.0
celery==3.1.23
Django==1.9
django-filter==0.14.0
django-mptt==0.8.5
django-mptt-admin==0.4.0
djangorestframework==3.4.4
dnspython==1.14.0
ipaddress==1.0.16
kombu==3.0.35
libsnmp==2.0.5
Markdown==2.6.6
mysql==0.0.1
MySQL-python==1.2.5
netaddr==0.7.18
netsnmpagent==0.5.2
nose==1.3.7
ply==3.8
prettytable==0.7.2
pyasn1==0.1.9
pycrypto==2.6.1
pysmi==0.0.7
pysnmp==4.3.2
pytz==2016.6.1
pyzabbix==0.7.4
redis==2.10.5
requests==2.11.1
six==1.10.0
vine==1.1.1
